chore(demo): restructure demo
this is to align better with how devs use better auth before they install this plugin
