+++ demo/src/app/(frontend)/dashboard/page.tsx
@@ -1,9 +1,8 @@
 import { headers } from "next/headers";
-import { redirect } from "next/navigation";
 import UserCard from "./user-card";
 import { OrganizationCard } from "./organization-card";
 import AccountSwitcher from "@/components/account-switch";
-import getPayload from "@/lib/payload";
+import { getPayload } from "@/lib/payload";
 import { cookies } from "next/headers";
 
 async function setSessionCookie(session: any) {
