+++ demo/src/payload/views/login/index.tsx
@@ -1,11 +1,8 @@
-import React from "react";
-import type { AdminViewServerProps, ServerProps } from "payload";
+import SignIn from "@/components/sign-in";
+import { Gutter } from "@payloadcms/ui";
 import { RenderServerComponent } from "@payloadcms/ui/elements/RenderServerComponent";
 import { redirect } from "next/navigation";
-import getPayload from "@/lib/payload";
-import { AuthForm } from "./form";
-import { Gutter } from "@payloadcms/ui";
-import SignIn from "@/components/sign-in";
+import type { AdminViewServerProps, ServerProps } from "payload";
 
 export default async function LoginView({
   initPageResult,
